refactor: Moved slot to single folder
