+++ @navikt/core/react/src/slot/merge-props.ts
@@ -1,37 +1,3 @@
-// https://github.com/radix-ui/primitives/blob/main/packages/react/slot/src/Slot.tsx
-import * as React from "react";
-import { mergeRefs } from "./hooks/useMergeRefs";
-
-interface SlotProps extends React.HTMLAttributes<HTMLElement> {
-  children?: React.ReactNode;
-}
-
-export const Slot = React.forwardRef<HTMLElement, SlotProps>(
-  (props, forwardedRef) => {
-    const { children, ...slotProps } = props;
-
-    if (React.isValidElement(children)) {
-      return React.cloneElement<any>(children, {
-        ...mergeProps(slotProps, children.props),
-        ref: forwardedRef
-          ? mergeRefs([forwardedRef, (children as any).ref])
-          : (children as any).ref,
-      });
-    }
-
-    if (React.Children.count(children) > 1) {
-      const error = new Error(
-        "Aksel: Components using 'asChild' expects to recieve a single React element child.",
-      );
-      error.name = "SlotError";
-      Error.captureStackTrace?.(error, Slot);
-      throw error;
-    }
-
-    return null;
-  },
-);
-
 function mergeProps(
   slotProps: Record<string, any>,
   childProps: Record<string, any>,
@@ -69,3 +35,5 @@ function mergeProps(
 
   return { ...slotProps, ...overrideProps };
 }
+
+export { mergeProps };
